Add type for return value

+++ src/initial-utils.jl
@@ -48,20 +48,32 @@ struct Co2PartitionCoeff{T}
                                        2.041e-3)
 end
 
-meq_silicic(Pw, Pc, Temp, b1, b2, b3, b4, b5, b6)                           = 1e-2*real((b1*Complex(Pw)^0.5+b2*Pw+b3*Complex(Pw)^1.5)/Temp+b4*Complex(Pw)^1.5+Pc*(b5*Complex(Pw)^0.5+b6*Pw))
-dmeqdT_silicic(Pw, Temp, b1, b2, b3)                                        = 1e-2*real(-1*(b1*Complex(Pw)^0.5+b2*Pw+b3*Complex(Pw)^1.5)*Temp^(-2))
-dmeqdP_silicic(Pw, dPwdP, Pc, dPcdP, Temp, b1, b2, b3, b4, b5, b6)          = 1e-8*real(dPwdP*((1/Temp)*(0.5*b1*Complex(Pw)^(-0.5)+b2+1.5*b3*Complex(Pw)^0.5)+1.5*b4*Complex(Pw)^0.5+Pc*(0.5*b5*Complex(Pw)^(-0.5)+b6))+dPcdP*(b5*Complex(Pw)^0.5+b6*Pw))
-dmeqdXco2_silicic(Pw, dPwdXco2, Pc, dPcdXco2, Temp, b1, b2, b3, b4, b5, b6) = 1e-2*real(dPwdXco2*((1/Temp)*(0.5*b1*Complex(Pw)^(-0.5)+b2+1.5*b3*Complex(Pw)^0.5)+1.5*b4*Complex(Pw)^0.5+Pc*(0.5*b5*Complex(Pw)^(-0.5)+b6))+dPcdXco2*(b5*Complex(Pw)^0.5+b6*Pw))
+meq_silicic(Pw::Float64, Pc::Float64, Temp::Float64, b1::Float64, b2::Float64, b3::Float64, b4::Float64, b5::Float64, b6::Float64)::Float64 = 
+    1e-2*real((b1*Complex(Pw)^0.5+b2*Pw+b3*Complex(Pw)^1.5)/Temp+b4*Complex(Pw)^1.5+Pc*(b5*Complex(Pw)^0.5+b6*Pw))
+dmeqdT_silicic(Pw::Float64, Temp::Float64, b1::Float64, b2::Float64, b3::Float64)::Float64 = 
+    1e-2*real(-1*(b1*Complex(Pw)^0.5+b2*Pw+b3*Complex(Pw)^1.5)*Temp^(-2))
+dmeqdP_silicic(Pw::Float64, dPwdP::Float64, Pc::Float64, dPcdP::Float64, Temp::Float64, b1::Float64, b2::Float64, b3::Float64, b4::Float64, b5::Float64, b6::Float64)::Float64 = 
+    1e-8*real(dPwdP*((1/Temp)*(0.5*b1*Complex(Pw)^(-0.5)+b2+1.5*b3*Complex(Pw)^0.5)+1.5*b4*Complex(Pw)^0.5+Pc*(0.5*b5*Complex(Pw)^(-0.5)+b6))+dPcdP*(b5*Complex(Pw)^0.5+b6*Pw))
+dmeqdXco2_silicic(Pw::Float64, dPwdXco2::Float64, Pc::Float64, dPcdXco2::Float64, Temp::Float64, b1::Float64, b2::Float64, b3::Float64, b4::Float64, b5::Float64, b6::Float64)::Float64 = 
+    1e-2*real(dPwdXco2*((1/Temp)*(0.5*b1*Complex(Pw)^(-0.5)+b2+1.5*b3*Complex(Pw)^0.5)+1.5*b4*Complex(Pw)^0.5+Pc*(0.5*b5*Complex(Pw)^(-0.5)+b6))+dPcdXco2*(b5*Complex(Pw)^0.5+b6*Pw))
 
-meq_mafic(P, T_C, X_co2, b1, b2, b3, b4, b5, b6, b7, b8, b9, b10) = 1e-2*real(b1+b2*T_C+b3*X_co2+b4*P+b5*T_C*X_co2+b6*T_C*P+b7*X_co2*P+b8*T_C^2+b9*Complex(X_co2)^2+b10*Complex(P)^2)
-dmeqdT_mafic(P, T_C, X_co2, b2, b5, b6, b8)                       = 1e-2*real(b2+b5*X_co2+b6*P+2*b8*T_C)
-dmeqdP_mafic(P, T_C, X_co2, b4, b6, b7, b10)                      = 1e-8*real(b4+b6*T_C+b7*X_co2+2*b10*P)
-dmeqdXco2_mafic(P, T_C, X_co2, b3, b5, b7, b9)                    = 1e-2*real(b3+b5*T_C+b7*P+2*b9*X_co2)
+meq_mafic(P::Float64, T_C::Float64, X_co2::Float64, b1::Float64, b2::Float64, b3::Float64, b4::Float64, b5::Float64, b6::Float64, b7::Float64, b8::Float64, b9::Float64, b10::Float64)::Float64 = 
+    1e-2*real(b1+b2*T_C+b3*X_co2+b4*P+b5*T_C*X_co2+b6*T_C*P+b7*X_co2*P+b8*T_C^2+b9*Complex(X_co2)^2+b10*Complex(P)^2)
+dmeqdT_mafic(P::Float64, T_C::Float64, X_co2::Float64, b2::Float64, b5::Float64, b6::Float64, b8::Float64)::Float64 = 
+    1e-2*real(b2+b5*X_co2+b6*P+2*b8*T_C)
+dmeqdP_mafic(P::Float64, T_C::Float64, X_co2::Float64, b4::Float64, b6::Float64, b7::Float64, b10::Float64)::Float64 = 
+    1e-8*real(b4+b6*T_C+b7*X_co2+2*b10*P)
+dmeqdXco2_mafic(P::Float64, T_C::Float64, X_co2::Float64, b3::Float64, b5::Float64, b7::Float64, b9::Float64)::Float64 = 
+    1e-2*real(b3+b5*T_C+b7*P+2*b9*X_co2)
 
-C_co2_f(Pw, Pc, Temp, c1, c2, c3, c4)                             = 1e-6*real(Pc*(c1+c2*Pw)/Temp+Pc*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5))
-dC_co2dT_f(Pw, Pc, Temp, c1, c2)                                  = 1e-6*real(-1*Pc*(c1+c2*Pw)*Complex(Temp)^(-2))
-dC_co2dP_f(Pw, Pc, Temp, dPwdP, dPcdP, c1, c2, c3, c4)            = 1e-12*real(Complex(Temp)^(-1)*(dPcdP*(c1+c2*Pw)+Pc*(c2*dPwdP))+dPcdP*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5)+Pc*(0.5*c3*Complex(Pw)^(-0.5)*dPwdP+1.5*c4*Complex(Pw)^0.5*dPwdP))
-dC_co2dXco2_f(Pw, Pc, Temp, dPwdXco2, dPcdXco2, c1, c2, c3, c4)   = 1e-6*real(Complex(Temp)^(-1)*(dPcdXco2*(c1+c2*Pw)+Pc*(c2*dPwdXco2))+dPcdXco2*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5)+Pc*(0.5*c3*Complex(Pw)^(-0.5)*dPwdXco2+1.5*c4*Complex(Pw)^0.5*dPwdXco2))
+C_co2_f(Pw::Float64, Pc::Float64, Temp::Float64, c1::Float64, c2::Float64, c3::Float64, c4::Float64)::Float64 = 
+    1e-6*real(Pc*(c1+c2*Pw)/Temp+Pc*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5))
+dC_co2dT_f(Pw::Float64, Pc::Float64, Temp::Float64, c1::Float64, c2::Float64)::Float64 = 
+    1e-6*real(-1*Pc*(c1+c2*Pw)*Complex(Temp)^(-2))
+dC_co2dP_f(Pw::Float64, Pc::Float64, Temp::Float64, dPwdP::Float64, dPcdP::Float64, c1::Float64, c2::Float64, c3::Float64, c4::Float64)::Float64 = 
+    1e-12*real(Complex(Temp)^(-1)*(dPcdP*(c1+c2*Pw)+Pc*(c2*dPwdP))+dPcdP*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5)+Pc*(0.5*c3*Complex(Pw)^(-0.5)*dPwdP+1.5*c4*Complex(Pw)^0.5*dPwdP))
+dC_co2dXco2_f(Pw::Float64, Pc::Float64, Temp::Float64, dPwdXco2::Float64, dPcdXco2::Float64, c1::Float64, c2::Float64, c3::Float64, c4::Float64)::Float64 = 
+    1e-6*real(Complex(Temp)^(-1)*(dPcdXco2*(c1+c2*Pw)+Pc*(c2*dPwdXco2))+dPcdXco2*(c3*Complex(Pw)^0.5+c4*Complex(Pw)^1.5)+Pc*(0.5*c3*Complex(Pw)^(-0.5)*dPwdXco2+1.5*c4*Complex(Pw)^0.5*dPwdXco2))
 
 function build_meq_silicic(Pw::T, Pc::T, Temp::T, dPwdP::T, dPcdP::T, dPwdXco2::T, dPcdXco2::T)::Vector{T} where T<:Float64
     @unpack b1, b2, b3, b4, b5, b6 = ExsolveSilicic()
